More generalized functionHandler with examples
