+++ examples/objectStorage.ts
@@ -1,6 +1,6 @@
-import { FunctionHandler, ObjectStorage } from "../src";
+import { Handler } from "../src";
 
-export const handler: FunctionHandler<ObjectStorage> = (event) => {
+export const handler: Handler.ObjectStorage = (event) => {
     return {
         statusCode: 200,
         body: event.messages[0].details.bucket_id,
