+++ examples/timer.ts
@@ -1,6 +1,6 @@
-import { FunctionHandler, Timer } from "../src";
+import { Handler } from "../src";
 
-export const handler: FunctionHandler<Timer> = (event) => {
+export const handler: Handler.Timer = (event) => {
     return {
         statusCode: 200,
         body: event.messages[0].event_metadata.event_type,
