+++ src/index.ts
@@ -1,17 +1,14 @@
-import type FunctionContext from "./context";
-export * from "./api-gateway";
-export * from "./http";
-export * from "./triggers";
+import { ApiGatewayAuthorizer } from "./api-gateway";
+import { FunctionHandler } from "./functionHandler";
+import { Http } from "./http";
+import { ObjectStorage, Timer } from "./triggers";
 
-type FunctionSyncHandler<Event, Result> = (
-    event: Event,
-    context: FunctionContext
-) => Result;
-type FunctionAsyncHandler<Event, Result> = (
-    event: Event,
-    context: FunctionContext
-) => Promise<Result>;
-
-export type FunctionHandler<Event, Result = unknown> =
-    | FunctionSyncHandler<Event, Result>
-    | FunctionAsyncHandler<Event, Result>;
+export namespace Handler {
+    export type ApiGatewayAuthorizer = FunctionHandler<
+        ApiGatewayAuthorizer.Event,
+        ApiGatewayAuthorizer.Result
+    >;
+    export type Http = FunctionHandler<Http.Event, Http.Result>;
+    export type ObjectStorage = FunctionHandler<ObjectStorage.Event>;
+    export type Timer = FunctionHandler<Timer.Event>;
+}
