+++ src/triggers/index.ts
@@ -4,14 +4,12 @@ import type ContainerRegistryExample from "./containerRegistry.json";
 import type DataStreamsExample from "./dataStreams.json";
 import type IotCoreExample from "./iotCore.json";
 import type MessageQueueExample from "./messageQueue.json";
-import type ObjectStorageExample from "./objectStorage.json";
-import type TimerExample from "./timer.json";
 
-export type Timer = typeof TimerExample;
 export type MessageQueue = typeof MessageQueueExample;
-export type ObjectStorage = typeof ObjectStorageExample;
 export type ContainerRegistry = typeof ContainerRegistryExample;
 export type CloudLogging = typeof CloudLoggingExample;
 export type IotCore = typeof IotCoreExample;
 export type Budget = typeof BudgetExample;
 export type DataStreams = typeof DataStreamsExample;
+export * from "./objectStorage";
+export * from "./timer";
